// src/orders/interfaces/order-w-products.interface.ts
import { OrderStatus } from "@prisma/client";

export interface OrderWithProducts {
  OrderItems: {
    name: string;
    productId: number;
    quantity: number;
    price: number;
  }[];
  id: string;
  totalAmount: number;
  totalItems: number;
  status: OrderStatus;
  paid: boolean;
  paidAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
}